+++ app/AppCache.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once __DIR__.'/AppKernel.php';
+namespace App;
 
 use Symfony\Bundle\FrameworkBundle\HttpCache\HttpCache;
