+++ app/console
@@ -8,8 +8,8 @@
 set_time_limit(0);
 
 require_once __DIR__.'/autoload.php';
-require_once __DIR__.'/AppKernel.php';
 
+use App\Kernel;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Component\Console\Input\ArgvInput;
 use Symfony\Component\Debug\Debug;
@@ -22,6 +22,6 @@ if ($debug) {
     Debug::enable();
 }
 
-$kernel = new AppKernel($env, $debug);
+$kernel = new Kernel($env, $debug);
 $application = new Application($kernel);
 $application->run($input);
