+++ cli-config.php
@@ -7,11 +7,11 @@
  */
 
 use Doctrine\ORM\Tools\Console\ConsoleRunner;
+use App\Kernel;
 
 require_once 'app/bootstrap.php.cache';
-require_once 'app/AppKernel.php';
 
-$kernel = new AppKernel('dev', true);
+$kernel = new Kernel('dev', true);
 $kernel->boot();
 $entityManager = $kernel->getContainer()->get('doctrine.orm.entity_manager');
