+++ composer.json
@@ -6,13 +6,19 @@
         "php": ">=7.2,<7.4",
         "claroline/distribution": "13.x-dev"
     },
+    "autoload": {
+        "psr-4": {
+            "App\\": "app/"
+        }
+    },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "2.14",
         "mockery/mockery": "0.9.9",
         "mikey179/vfsstream": "1.6.4",
         "phpmd/phpmd": "@stable",
         "phpunit/phpunit": "7.1.5",
-        "symfony/phpunit-bridge": "4.0.9"
+        "symfony/phpunit-bridge": "^4.0",
+        "stefk/md": "dev-master"
     },
     "scripts": {
         "fast-install" : [
